% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brief.R
\name{brief}
\alias{brief}
\title{data profiling}
\usage{
brief(inputData)
}
\arguments{
\item{inputData}{data frame}
}
\value{
data frame with changed column
}
\description{
dimension, missing, cardinality, correlation, zero, uniform, unique for each variables
}
\examples{
brief(iris)

}
