% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{summary statistics from data with character value}
\usage{
describe(i)
}
\arguments{
\item{i}{data}
}
\value{
named list
}
\description{
count, mean, sd, quantile values.
}
\examples{
describe(iris[, "Sepal.Width"])
describe(iris$Sepal.Length) # this will be work too.

}
\seealso{
base \code{summary} function
}
