% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{distribute}
\alias{distribute}
\title{see data's distribution.}
\usage{
distribute(i, xlabel = NULL)
}
\arguments{
\item{i}{vector of numeric data}

\item{xlabel}{label for x axis (default is \code{value}), y axis set as 'count'.}
}
\value{
histogram built with ggplot
}
\description{
build histogram from character input
}
\examples{
distribute(iris[,"Sepal.Width"])
distribute(iris$Sepal.Length) # this will be work too.


}
\seealso{
\code{geom_histogram}
}
