% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominate.R
\name{dominate}
\alias{dominate}
\title{find most common value in data}
\usage{
dominate(i)
}
\arguments{
\item{i}{data}
}
\value{
named list
}
\description{
most common value and count & percent, among howmuch values
}
\examples{
dominate(iris[, "Sepal.Length"])
scissor::binarize(iris, "Species", "In", 'c("virginica", "setosa")')

}
