% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distributionModule.R
\name{mod_distributionModule_server}
\alias{mod_distributionModule_server}
\title{Shiny Module for see distribution of variable.}
\usage{
mod_distributionModule_server(id, inputData)
}
\arguments{
\item{id}{id of module}

\item{inputData}{"reactive" data}
}
\description{
distributionModule Server Functions
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    mod_distributionModule_ui("module")
  )
  server <- function(input, output, session) {
    mod_distributionModule_server("module", reactive({
      iris
    }))
  }
  shinyApp(ui, server)
}

}
\seealso{
mod_distributionModule_ui of \code{board}
}
