% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distributionModule.R
\name{mod_distributionModule_ui}
\alias{mod_distributionModule_ui}
\title{Shiny Module for see distribution of variable.}
\usage{
mod_distributionModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
distributionModule UI Function.
}
\examples{
if(interactive()){
library(shiny)
ui <- fluidPage(
  mod_distributionModule_ui('module')
)
server <- function(input, output, session) {
  mod_distributionModule_server('module', reactive({iris}))
}
shinyApp(ui, server)
}

}
\seealso{
mod_distributionModule_server of \code{board}
}
