% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Outlier of data}
\usage{
outlier(i)
}
\arguments{
\item{i}{data}
}
\value{
named list
}
\description{
calculate outlier
}
\examples{
outlier(iris[,'Sepal.Width'])

}
\seealso{
https://en.wikipedia.org/wiki/Interquartile_range (IQR)
}
